requirements.txt
cryptography~=41.0.5
pycryptodome~=3.19.0
